Rename `operator-config` to `install-config` Package `operatorconfig` -> `installconfig` ConfigMap `operator-config` -> `install-config`
