+++ pkg/api/validation/edgeconnect/module_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/Dynatrace/dynatrace-operator/pkg/util/operatorconfig"
+	"github.com/Dynatrace/dynatrace-operator/pkg/util/installconfig"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -12,12 +12,12 @@ func TestIsModuleDisabled(t *testing.T) {
 	ctx := context.Background()
 
 	t.Run("module disabled => error", func(t *testing.T) {
-		errMsg := isModuleDisabled(ctx, &Validator{modules: operatorconfig.Modules{EdgeConnect: false}}, nil)
+		errMsg := isModuleDisabled(ctx, &Validator{modules: installconfig.Modules{EdgeConnect: false}}, nil)
 		assert.Equal(t, errorModuleDisabled, errMsg)
 	})
 
 	t.Run("module enabled => no error", func(t *testing.T) {
-		errMsg := isModuleDisabled(ctx, &Validator{modules: operatorconfig.Modules{EdgeConnect: true}}, nil)
+		errMsg := isModuleDisabled(ctx, &Validator{modules: installconfig.Modules{EdgeConnect: true}}, nil)
 		assert.Empty(t, errMsg)
 	})
 }
